moving the chat.ts to a file and adding a model for the chats
