+++ server/src/sockets/chat.ts
@@ -1,7 +1,8 @@
 import { Server } from "socket.io";
 import http from "http";
-import { dbClient } from "./models/db";
+import { dbClient } from "../models/db";
 import { randomUUID } from "crypto";
+import { saveMessageToDb } from "../models/chat";
 
 export function setupChatSocket(server: http.Server) {
     const io = new Server(server, {
@@ -15,13 +16,10 @@ export function setupChatSocket(server: http.Server) {
         console.log("Client connected:", socket.id);
 
         socket.on("message", async (data) => {
-            console.log("Message receivrd:", data);
+            console.log("Message received:", data);
 
             try { 
-                await dbClient.execute({
-                    sql: `INSERT INTO messages (id, chat_id, sender_id, receiver_id, text, created_at) VALUES (?, ?, ?, ?, ?, ?)`,
-                    args: [randomUUID(), data.chat_id, data.sender_id, data.receiver_id, data.text, new Date().toISOString()]
-                });    
+                await saveMessageToDb(data);
             } catch (error) {
                 console.error("Failed to insert message:", error);
             }
